update branch events
