--- frontend/app/src/entities/events/ui/branch-events/global-branch-event.tsx
@@ -1,22 +1,0 @@
-import { NodeLabel } from "@/entities/nodes/object/ui/node-label";
-import { EventNodeInterface } from "@/shared/api/graphql/generated/graphql";
-import { Icon } from "@iconify-icon/react";
-import { BRANCH_EVENTS_MAPPING } from "./branch-event";
-
-export const BranchEvent = (props: EventNodeInterface) => {
-  const { event, account_id } = props;
-
-  return (
-    <div className="flex items-center justify-between">
-      <div className="flex items-center gap-2 text-sm text-gray-500">
-        <Icon icon="mdi:source-branch" className="text-gray-400" />
-
-        <div className="text-black font-semibold">
-          <NodeLabel id={account_id} />
-        </div>
-
-        {(BRANCH_EVENTS_MAPPING[event] && BRANCH_EVENTS_MAPPING[event](props)) ?? event}
-      </div>
-    </div>
-  );
-};
